feat: Add sign-in and registration error handling
Error handling has been added to the sign-in and registration functions in the `AuthProvider` class. If there is an error during sign-in or registration, a snackbar with the error message will be displayed to the user. This improves the user experience by providing feedback when there are issues with authentication.