feat: Add onboarding screen
The code changes include importing the `OnBoarding` screen and adding a case for `/onboarding` in the `Routes` class. This is in line with recent user commits that mention a homepage and list report redesign, as well as store report and edit UI for login and registration. The recent repository commits also include updating dependencies and fixing package versions, as well as updating the `flutter_onboarding_slider` package.